requirements.txt
asgiref==3.5.2
boto3==1.21.45
botocore==1.24.45
cffi==1.15.1
colorama==0.4.5
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.10.0
django-credit-cards==0.4.1
django-database-url==1.0.3
django-mirage-field==1.4.0
django-simple-history==3.0.0
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.5
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.10